Fix test and improve code
