+++ src/dll.c
@@ -19,6 +19,7 @@
  */
 
 #include <math.h>
+#include <stdlib.h>
 #include "utils.h"
 #include "dll.h"
